Fix invalid directory error (#334)
We cd to a temporary directory at the beginning of every Airlpane task.
We do not switch back to another directory at the end of the task. At
the end of an Airplane run, this doesn't matter, because Airplane only
runs 1 task during one execution. But unit tests fail with this
behavior. This change makes functions that depend on a directory
existing in order to change back to it to only conditionally change back
to that directory - only if it existed.