Keep running workflow is account closure fails (#523)
AWS has a limit on the number of accounts it can close per month - I
think it is based as a percentage of total accounts you own. When we
delete a lot of AWS accounts, this can prevent us from closing some
until next month. When this happens, this should not stop running other
teardown things on other accounts, like DNS removal or mage teardowns.
This change makes the workflow keep running.